+++ src/main/java/org/monarchinitiative/phenotefx/gui/editrow/EditRowView.java
@@ -1,0 +1,6 @@
+package org.monarchinitiative.phenotefx.gui.editrow;
+
+import org.monarchinitiative.phenotefx.framework.FXMLView;
+
+public class EditRowView  extends FXMLView {
+}