+++ src/main/java/org/monarchinitiative/phenotefx/gui/main/PhenoteView.java
@@ -1,4 +1,4 @@
-package org.monarchinitiative.phenotefx.gui.phenote;
+package org.monarchinitiative.phenotefx.gui.main;
 
 /*
  * #%L