+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # PhenoteFX Changelog
 
+## v0.2.3
+* Read-the-docs
+* context menu to change evidence code
+* refactor to use phenol instead of OLPG
+* dialog to update the publication (citation)
+
+
 ## v0.2.2
 * ingest local copy of hp.obo file